refactor: MenuGroup 리팩토링
